requirements.txt
anthropic>=0.40.0
mcp>=1.14.0
fastapi>=0.110
uvicorn[standard]>=0.30
httpx>=0.27
pyyaml>=6.0.1
python-dotenv>=1.0
pytest>=8.3
pytest-asyncio>=0.23
coverage>=7.6
PyJWT>=2.8.0
cryptography>=41.0.0
